Splitted functions and classes from the script
